+++ frontend/src/App.jsx
@@ -8,7 +8,7 @@ import Login from './pages/login'
 import Register from './pages/register'
 import Problemset from './pages/problemset'
 import Sidebar from './components/Sidebar'
-import List from './pages/list'
+// import List from './pages/problemset'
 import QuestionDetails from './pages/QuestionDetails'
 
 function Home() {
@@ -33,7 +33,6 @@ function App() {
           <Route path="/login" element={<Login />} />
           <Route path="/register" element={<Register />} />
           <Route path="/problemset" element={<Problemset />} />
-          <Route path="/list" element={<List />} />
           <Route path="/question/:qid" element={<QuestionDetails />} />
         </Routes>
       </main>
