+++ frontend/src/components/Navbar.jsx
@@ -33,15 +33,15 @@ function Navbar(){
                       <li>
                            <Link to="/problemset" className="block py-2 px-3 text-gray-900 rounded-sm hover:bg-gray-100 md:hover:bg-transparent md:border-0 md:hover:text-blue-700 md:p-0 dark:text-white md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent">Problemset</Link>
                       </li>
-                      <li>
+                      {/* <li>
                            <Link to="/list" className="block py-2 px-3 text-gray-900 rounded-sm hover:bg-gray-100 md:hover:bg-transparent md:border-0 md:hover:text-blue-700 md:p-0 dark:text-white md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent">List</Link>
-                      </li>
-                      <li>
+                      </li> */}
+                      {/* <li>
                           <Link to="/contribute" className="block py-2 px-3 text-gray-900 rounded-sm hover:bg-gray-100 md:hover:bg-transparent md:border-0 md:hover:text-blue-700 md:p-0 dark:text-white md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent">Contribute</Link>
-                      </li>
-                      <li>
+                      </li> */}
+                      {/* <li>
                            <Link to="/report" className="block py-2 px-3 text-gray-900 rounded-sm hover:bg-gray-100 md:hover:bg-transparent md:border-0 md:hover:text-blue-700 md:p-0 dark:text-white md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent">Report</Link>
-                      </li>
+                      </li> */}
                       {!isLoggedIn ? (
               <>
                 <li>
